// src/app/models/ganancia.ts
export class Ganancia{
    venta_directa: number;
    fecha_cobro: Date;
    primera_linea: number;
    segunda_linea: number;
    tercera_linea: number;
    cuarta_linea: number;
    binario: number;
    total: number;
    puntos_izquierda: number;
    puntos_derecha: number;
    semanal: number;

    constructor(){}
}